Corrected values of mph2mps
Previously the wrong value was used in the mph2mps conversion. It was 2.24 but actually it is 0.44704. The confusion comes from the Udacity’s video incorreclty said the “mph in mps is 2.24”. What Aaron from the Udacity’s video meant to say is 5mph in mps is 2.24.

Correction is made throughout the code.