Added speed control logic maintaining safe gap to target vehicle
Replaced the reference velocity value assignment logic with`too_close` flag when the ego car is too close to the target vehicle and it is less than the safe gap distance (i.e. 30 meters).

The initial reference (ego) car velocity is set to 0 mph and then it will increase by increment of 0.5 mph until maximum velocity (i.e. 49.5 mph) is reached unless it is too close to the target car.

When the ego car is getting too close to the target car, the ego car will decrease the velocity by 0.5mph until the safe gap distance (i.e. 30 meters) is reached.