Enabled the prepare to change lane logic
Better results. Significantly less flickering when changing lanes. Able to achieve 5.22 miles without incident at first run!

First incident is collision when the ego vehicle was on the outermost right lane and there are two cars in the middle and the right lane. It got confused whether to choose to go middle and right lane and keep driving then collide with both vehicles.